cnf: allow to install found package
This streamlines the most common use-case I have:
"I want program foo ASAP and dont care how the package is called"

With the environment variable, this can conveniently be added to .bashrc
and we do not have to parse, document and maintain a config file.

Using subprocess.call to prevent shell code injection attacks.